requirements.txt
cassandra-driver==3.15.1
